requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cleo==0.8.1
click==8.0.3
clikit==0.6.2
colorama==0.4.4
crashtest==0.3.1
distlib==0.3.4
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.73.0
filelock==3.4.2
greenlet==1.1.2
gTTS==2.2.3
gunicorn==20.1.0
h11==0.13.0
heroku==0.1.4
html5lib==1.1
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
keyring==21.8.0
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==1.0.3
orjson==3.6.6
packaging==20.9
passlib==1.7.4
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.4.1
poetry==1.1.12
poetry-core==1.0.7
psycopg2==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pylev==1.4.0
pyparsing==3.0.7
python-dateutil==1.5
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
sqlalchemy2-stubs==0.0.2a20
sqlmodel==0.0.6
starlette==0.17.1
tomlkit==0.9.1
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
virtualenv==20.13.1
watchgod==0.7
webencodings==0.5.1
websockets==10.1

pytest~=7.1.2